Do ignore routes event if generated during build
Even if a route is explicitly marked as ignored by Stenope in its
routing options, if the route was used to generate an URL inside a
non-ignored route, it'll still be part of the build.

This is problematic, even more when the route is only used to generate
an URL to an external resource.

Let's properly ignore such routes in the UrlGenerator.